features: Added Authentification + Joi validation + Finished Candidate routes

+++ controllers/candidate.controller.js
@@ -1,5 +1,5 @@
 import getPool from "../databases/pool.js";
-import { putStringifier } from "../functions/routeFunc.js";
+import { findXTarget, putStringifier } from "../functions/routeFunc.js";
 
 const getAllCandidates = async (_, res) => {
   try {
@@ -20,6 +20,28 @@ const getAllCandidates = async (_, res) => {
 const getUserCandidate = async (req, res) => {
   try {
     const pool = await getPool();
+    const target = Number(req.params.id);
+    const found = await findXTarget(target);
+    if (found) {
+      const userListOfMissions = await pool.query(
+        `
+        SELECT c.id_candidate, c.status, m.id_mission, m.title, m.description, m.date
+        FROM candidate c
+        JOIN mission m ON c.fk_mission = m.id_mission
+        WHERE c.fk_user = ?;`,
+        target
+      );
+      if (userListOfMissions) {
+        res.status(200).json({
+          message: "All missions this user Candidated to:",
+          missions: userListOfMissions[0],
+        });
+      } else {
+        res.status(500).json({ message: "Failed to show all . . ." });
+      }
+    } else {
+      res.status(404).json({ message: "User doesn't exist" });
+    }
   } catch (err) {
     console.error(err);
     res.status(500).json({ message: "Something went wrong while " });
@@ -29,6 +51,29 @@ const getUserCandidate = async (req, res) => {
 const getMissionCandidate = async (req, res) => {
   try {
     const pool = await getPool();
+    const target = Number(req.params.id);
+    const found = await findXTarget(target, "mission");
+    if (found) {
+      const candidatesForThisMission = await pool.query(
+        `
+        SELECT c.id_candidate, c.status, m.title, u.id_user, u.name, u.email
+        FROM candidate c
+        JOIN users u ON c.fk_user = u.id_user
+        JOIN mission m ON c.fk_mission = m.id_mission
+        WHERE c.fk_mission = ?;`,
+        target
+      );
+      if (candidatesForThisMission) {
+        res.status(200).json({
+          message: `List of all candidates of mission.id_mission=${target}`,
+          candidates: candidatesForThisMission[0],
+        });
+      } else {
+        res.status(500).json({ message: "Failed to show all . . ." });
+      }
+    } else {
+      res.status(404).json({ message: "Mission doesn't exist" });
+    }
   } catch (err) {
     console.error(err);
     res.status(500).json({ message: "Something went wrong while " });
@@ -63,6 +108,27 @@ const postCandidate = async (req, res) => {
 const candidateStatus = async (req, res) => {
   try {
     const pool = await getPool();
+    const target = Number(req.params.id);
+    const found = await findXTarget(target, "candidate");
+    if (found) {
+      const { status } = req.body;
+      const jsonValues = {
+        status: status,
+      };
+      const transformedData = putStringifier(jsonValues);
+      const command = `UPDATE \`candidate\` SET ${transformedData[0]} WHERE candidate.id_candidate = ${target};`;
+      const modifyCandidate = await pool.query(command, transformedData[1]);
+      if (modifyCandidate) {
+        // prettier-ignore
+        res.status(200).json({ message: "Modified candidate's status successfully" });
+      } else {
+        res
+          .status(400)
+          .json({ message: "Modifying candidate's status failed" });
+      }
+    } else {
+      res.status(404).json({ message: "Candidate doesn't exist" });
+    }
   } catch (err) {
     console.error(err);
     // prettier-ignore
@@ -74,14 +140,19 @@ const deleteCandidate = async (req, res) => {
   try {
     const pool = await getPool();
     const target = Number(req.params.id);
-    const deleteCandidate = await pool.query(`
+    const found = await findXTarget(target, "candidate");
+    if (found) {
+      const deleteCandidate = await pool.query(`
       DELETE FROM candidate WHERE candidate.id_candidate = ${target}`);
-    if (deleteCandidate) {
-      res.status(204).json({
-        message: `Successfully delete candidate id_candidate=${target}`,
-      });
+      if (deleteCandidate) {
+        res.status(204).json({
+          message: `Successfully delete candidate id_candidate=${target}`,
+        });
+      } else {
+        res.status(500).json({ message: "Failed to delete candidate" });
+      }
     } else {
-      res.status(500).json({ message: "Failed to delete candidate" });
+      res.status(404).json({ message: "Candidate doesn't exist" });
     }
   } catch (err) {
     console.error(err);
